Create Page helper class for Cypress
