+++ cypress/e2e/charts/table.cy.js
@@ -1,6 +1,7 @@
-import { tableCypherQuery } from '../fixtures/cypher_queries';
+import { tableCypherQuery } from '../../fixtures/cypher_queries';
+import { Page } from '../../Page';
 
-const WAITING_TIME = 20000;
+const page = new Page();
 // Ignore warnings that may appear when using the Cypress dev server
 Cypress.on('uncaught:exception', (err, runnable) => {
   console.log(err, runnable);
@@ -9,30 +10,7 @@ Cypress.on('uncaught:exception', (err, runnable) => {
 
 describe('Testing table', () => {
   beforeEach('open neodash', () => {
-    cy.viewport(1920, 1080);
-    cy.visit('/', {
-      onBeforeLoad(win) {
-        win.localStorage.clear();
-      },
-    });
-
-    cy.get('#form-dialog-title', { timeout: 20000 }).should('contain', 'NeoDash - Neo4j Dashboard Builder').click();
-
-    cy.get('#form-dialog-title').then(($div) => {
-      const text = $div.text();
-      if (text == 'NeoDash - Neo4j Dashboard Builder') {
-        cy.wait(500);
-        // Create new dashboard
-        cy.contains('New Dashboard').click();
-      }
-    });
-
-    cy.get('#form-dialog-title', { timeout: 20000 }).should('contain', 'Connect to Neo4j');
-
-    cy.get('#url').clear().type('localhost');
-    cy.get('#dbusername').clear().type('neo4j');
-    cy.get('#dbpassword').type('test1234');
-    cy.get('button').contains('Connect').click();
+    page.init().createNewDashboard().connectToNeo4j();
     cy.wait(100);
   });
