+++ cypress/e2e/start_page.cy.js
@@ -10,7 +10,11 @@ import {
   gaugeChartCypherQuery,
   formCypherQuery,
 } from '../fixtures/cypher_queries';
-import { createReportOfType, selectReportOfType, enableAdvancedVisualizations, enableFormsExtension } from './utils';
+
+import { Page } from '../Page';
+
+const CARD_SELECTOR = 'main .react-grid-item:eq(2)';
+const page = new Page(CARD_SELECTOR);
 
 const WAITING_TIME = 20000;
 // Ignore warnings that may appear when using the Cypress dev server
@@ -21,40 +25,7 @@ Cypress.on('uncaught:exception', (err, runnable) => {
 
 describe('NeoDash E2E Tests', () => {
   beforeEach(() => {
-    cy.viewport(1920, 1080);
-    // Navigate to index
-    cy.visit('/', {
-      onBeforeLoad(win) {
-        win.localStorage.clear();
-      },
-    });
-
-    cy.get('#form-dialog-title', { timeout: 20000 }).should('contain', 'NeoDash - Neo4j Dashboard Builder').click();
-
-    cy.get('#form-dialog-title').then(($div) => {
-      const text = $div.text();
-      if (text == 'NeoDash - Neo4j Dashboard Builder') {
-        cy.wait(500);
-        // Create new dashboard
-        cy.contains('New Dashboard').click();
-      }
-    });
-
-    // If an old dashboard exists in cache, do a check to make sure we clear it.
-    // if (cy.contains("Create new dashboard")) {
-    //     cy.contains('Yes').click()
-    // }
-
-    cy.get('#form-dialog-title', { timeout: 20000 }).should('contain', 'Connect to Neo4j');
-
-    // Connect to Neo4j database
-    // cy.get('#protocol').click()
-    // cy.contains('neo4j').click()
-    cy.get('#url').clear().type('localhost');
-    // cy.get('#database').type('neo4j')
-    cy.get('#dbusername').clear().type('neo4j');
-    cy.get('#dbpassword').type('test1234');
-    cy.get('button').contains('Connect').click();
+    page.init().createNewDashboard().connectToNeo4j();
     cy.wait(100);
   });
 
@@ -98,7 +69,7 @@ describe('NeoDash E2E Tests', () => {
 
   it('creates a bar chart report', () => {
     cy.checkInitialState();
-    createReportOfType('Bar Chart', barChartCypherQuery);
+    page.createReportOfType('Bar Chart', barChartCypherQuery);
     cy.get('main .react-grid-item:eq(2) #index input[name="Category"]', { timeout: WAITING_TIME }).should(
       'have.value',
       'released'
@@ -109,7 +80,7 @@ describe('NeoDash E2E Tests', () => {
 
   it('creates a pie chart report', () => {
     cy.checkInitialState();
-    createReportOfType('Pie Chart', barChartCypherQuery);
+    page.createReportOfType('Pie Chart', barChartCypherQuery);
     cy.get('main .react-grid-item:eq(2) #index input[name="Category"]', { timeout: WAITING_TIME }).should(
       'have.value',
       'released'
@@ -121,7 +92,7 @@ describe('NeoDash E2E Tests', () => {
 
   it('creates a line chart report', () => {
     cy.checkInitialState();
-    createReportOfType('Line Chart', barChartCypherQuery);
+    page.createReportOfType('Line Chart', barChartCypherQuery);
     cy.get('main .react-grid-item:eq(2) #x input[name="X-value"]', { timeout: WAITING_TIME }).should(
       'have.value',
       'released'
@@ -136,7 +107,7 @@ describe('NeoDash E2E Tests', () => {
 
   it('creates a map chart report', () => {
     cy.checkInitialState();
-    createReportOfType('Map', mapChartCypherQuery, true);
+    page.createReportOfType('Map', mapChartCypherQuery, true);
     cy.get('main .react-grid-item:eq(2) .MuiCardContent-root svg > g > path', { timeout: WAITING_TIME }).should(
       'have.length',
       5
@@ -145,7 +116,7 @@ describe('NeoDash E2E Tests', () => {
 
   it('creates a single value report', () => {
     cy.checkInitialState();
-    createReportOfType('Single Value', barChartCypherQuery);
+    page.createReportOfType('Single Value', barChartCypherQuery);
     cy.get('main .react-grid-item:eq(2) .MuiCardContent-root > div > div:nth-child(2) > span', {
       timeout: WAITING_TIME,
     })
@@ -156,16 +127,16 @@ describe('NeoDash E2E Tests', () => {
   });
 
   it.skip('creates a gauge chart report', () => {
-    enableAdvancedVisualizations();
+    page.enableAdvancedVisualizations();
     cy.checkInitialState();
-    createReportOfType('Gauge Chart', gaugeChartCypherQuery);
+    page.createReportOfType('Gauge Chart', gaugeChartCypherQuery);
     cy.get('.text-group > text', { timeout: WAITING_TIME }).contains('69');
   });
 
   it('creates a sunburst chart report', () => {
-    enableAdvancedVisualizations();
+    page.enableAdvancedVisualizations();
     cy.checkInitialState();
-    createReportOfType('Sunburst Chart', sunburstChartCypherQuery);
+    page.createReportOfType('Sunburst Chart', sunburstChartCypherQuery);
     cy.get('main .react-grid-item:eq(2) #index input[name="Path"]', { timeout: WAITING_TIME }).should(
       'have.value',
       'x.path'
@@ -175,9 +146,9 @@ describe('NeoDash E2E Tests', () => {
   });
 
   it('creates a circle packing report', () => {
-    enableAdvancedVisualizations();
+    page.enableAdvancedVisualizations();
     cy.checkInitialState();
-    createReportOfType('Circle Packing', sunburstChartCypherQuery);
+    page.createReportOfType('Circle Packing', sunburstChartCypherQuery);
     cy.get('main .react-grid-item:eq(2) #index input[name="Path"]', { timeout: WAITING_TIME }).should(
       'have.value',
       'x.path'
@@ -187,9 +158,9 @@ describe('NeoDash E2E Tests', () => {
   });
 
   it('creates a tree map report', () => {
-    enableAdvancedVisualizations();
+    page.enableAdvancedVisualizations();
     cy.checkInitialState();
-    createReportOfType('Treemap', sunburstChartCypherQuery);
+    page.createReportOfType('Treemap', sunburstChartCypherQuery);
     cy.get('main .react-grid-item:eq(2) #index input[name="Path"]', { timeout: WAITING_TIME }).should(
       'have.value',
       'x.path'
@@ -199,9 +170,9 @@ describe('NeoDash E2E Tests', () => {
   });
 
   it('creates a sankey chart report', () => {
-    enableAdvancedVisualizations();
+    page.enableAdvancedVisualizations();
     cy.checkInitialState();
-    createReportOfType('Sankey Chart', sankeyChartCypherQuery, true);
+    page.createReportOfType('Sankey Chart', sankeyChartCypherQuery, true);
     cy.get('main .react-grid-item:eq(2) .MuiCardContent-root svg > g > path', { timeout: WAITING_TIME }).should(
       'have.attr',
       'fill-opacity',
@@ -211,7 +182,7 @@ describe('NeoDash E2E Tests', () => {
 
   it('creates a raw json report', () => {
     cy.checkInitialState();
-    createReportOfType('Raw JSON', barChartCypherQuery);
+    page.createReportOfType('Raw JSON', barChartCypherQuery);
     cy.get('main .react-grid-item:eq(2) .MuiCardContent-root textarea:nth-child(1)', { timeout: 45000 }).should(
       ($div) => {
         const text = $div.text();
@@ -222,7 +193,7 @@ describe('NeoDash E2E Tests', () => {
 
   it('creates a parameter select report', () => {
     cy.checkInitialState();
-    selectReportOfType('Parameter Select');
+    page.selectReportOfType('Parameter Select');
     cy.wait(500);
     cy.get('#autocomplete-label-type').type('Movie');
     cy.get('#autocomplete-label-type-option-0').click();
@@ -236,20 +207,20 @@ describe('NeoDash E2E Tests', () => {
 
   it('creates an iframe report', () => {
     cy.checkInitialState();
-    createReportOfType('iFrame', iFrameText);
+    page.createReportOfType('iFrame', iFrameText);
     cy.get('main .react-grid-item:eq(2) .MuiCardContent-root iframe', { timeout: 45000 }).should('be.visible');
   });
 
   it('creates a markdown report', () => {
     cy.checkInitialState();
-    createReportOfType('Markdown', markdownText);
+    page.createReportOfType('Markdown', markdownText);
     cy.get('main .react-grid-item:eq(2) .MuiCardContent-root h1', { timeout: 45000 }).should('have.text', 'Hello');
   });
 
   it.skip('creates a form report', () => {
-    enableFormsExtension();
+    page.enableFormsExtension();
     cy.checkInitialState();
-    createReportOfType('Form', formCypherQuery, true, false);
+    page.createReportOfType('Form', formCypherQuery, true, false);
     cy.get('main .react-grid-item:eq(2) .form-add-parameter').click();
     cy.wait(200);
     cy.get('#autocomplete-label-type').type('Movie');
@@ -273,7 +244,7 @@ describe('NeoDash E2E Tests', () => {
   // TODO - this test is flaky, especially in GitHub actions environment.
   it.skip('test load dashboard from file and stress test report customizations', () => {
     try {
-      var NUMBER_OF_PAGES_IN_STRESS_TEST_DASHBOARD = 5;
+      const NUMBER_OF_PAGES_IN_STRESS_TEST_DASHBOARD = 5;
       const file = cy.request(loadDashboardURL).should((response) => {
         cy.get('#root .MuiDrawer-root .MuiIconButton-root:eq(2)').click();
         cy.get('.MuiDialog-root .MuiPaper-root .MuiDialogContent-root textarea:eq(0)')
